File communication interface.
Allow qz-tray to read/write from files.
Writing to a file will either overwrite or append to contents of a file
depending on configuration.
Reading a file involves registering for notifications of file
modifications and recieving file contents in notification of file
modification.

For simplicity of implementation a FileWatcher is created for each File
that modification notifications are wanted for.